A modest attempt to appease the elaboration errors on nexys4ddr synthesis. - Elaboration step completes, but it dies during placement, with:
ERROR: [Place 30-157] Sub-optimal placement for an MMCM-MMCM cascade pair. If this sub optimal condition is acceptable for this design, you may use the CLOCK_DEDICATED_ROUTE constraint in the .xdc file to demote this message to a WARNING. However, the use of this override is highly discouraged. These examples can be used directly in the .xdc file to override this clock rule.